requirements.txt
altair==5.5.0
APScheduler==3.11.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.6
gevent==25.9.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.2.15
MarkupSafe==3.0.3
narwhals==2.7.0
numpy==2.2.6
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tomli==2.2.1
tornado==6.5.2
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.9.0
zope.event==6.0
zope.interface==8.0.1